Utils.java file added
