+++ library/src/main/resources/com/library/reset.fxml
@@ -7,9 +7,12 @@
 <?import javafx.scene.control.TextArea?>
 <?import javafx.scene.control.TextField?>
 <?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.StackPane?>
 <?import javafx.scene.text.Font?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="450.0" xmlns="http://javafx.com/javafx/22" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.library.ControllerReset">
+<StackPane fx:id="root" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="450.0" xmlns="http://javafx.com/javafx/23.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.library.ControllerReset">
+   <children>
+      <AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="450.0" >
    <children>
       <Label fx:id="label" layoutX="102.0" layoutY="85.0" text="Reset your password">
          <font>
@@ -38,3 +41,6 @@
       <Button fx:id="Cancel" layoutX="177.0" layoutY="546.0" mnemonicParsing="false" onAction="#CancelMethod" prefHeight="26.0" prefWidth="97.0" text="Cancel" />
    </children>
 </AnchorPane>
+   </children>
+</StackPane>
+
